backend/requirements.txt
fastapi==0.121.3
uvicorn[standard]==0.38.0
python-dotenv==1.2.1
pydantic==2.12.4
pydantic-settings==2.12.0
supabase==2.24.0
qdrant-client==1.16.0
google-generativeai==0.8.5
pdfplumber==0.11.8
httpx>=0.26.0,<0.29.0
python-multipart==0.0.20
slowapi==0.1.9
PyJWT==2.10.1